evaluation-observe/deploy-langfuse-on-ecs-with-fargate/langfuse-v2/requirements.txt
aws-cdk-lib==2.172.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment==3.0.82
